# backend/cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/teams": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[Cache: 5sec] Returns a list of all registered teams \u0026 their members. 🔒 Requires role: **admin**",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Get all teams",
                "responses": {
                    "200": {
                        "description": "List of teams retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TeamReadAll"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new team with the provided details. You can add members or not. 🔒 Requires role: **admin**",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Create a new team",
                "parameters": [
                    {
                        "description": "Team to create",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TeamCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Team created successfully"
                    }
                }
            }
        },
        "/teams/add-users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds users to an existing team by their UUIDs. 🔒 Requires role: **admin**",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Add users to a team",
                "parameters": [
                    {
                        "description": "Team UUID and User UUIDs to add",
                        "name": "team_users",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TeamAddUsers"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Users added to team successfully"
                    }
                }
            }
        },
        "/teams/edit/{uuid}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing team's details by its UUID. 🔒 Requires role: **admin**",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Update team by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated team details",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TeamUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team updated successfully"
                    }
                }
            }
        },
        "/teams/users/{team_uuid}/{user_uuid}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a user from a team by the provided team UUID and user UUID. 🔒 Requires role: **admin**",
                "tags": [
                    "Teams"
                ],
                "summary": "Remove user from team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team UUID",
                        "name": "team_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "user_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User removed from team successfully"
                    }
                }
            }
        },
        "/teams/{team_uuid}/users/{user_uuid}/edit-manager-status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a team member's manager status by team UUID and user UUID. 🔒 Requires role: **admin**",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Update team member manager status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team UUID",
                        "name": "team_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "user_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Is Manager Status (1 = true, 0 = false)",
                        "name": "is_manager",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    }
                }
            }
        },
        "/teams/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a team and its members by the provided UUID. 🔒 Requires role: **any**",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Get team by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/model.TeamReadAll"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a team by the provided UUID. 🔒 Requires role: **admin**",
                "tags": [
                    "Teams"
                ],
                "summary": "Delete team by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Team deleted successfully"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.NewTeamMember": {
            "type": "object",
            "required": [
                "user_uuid"
            ],
            "properties": {
                "is_manager": {
                    "type": "boolean"
                },
                "user_uuid": {
                    "type": "string"
                }
            }
        },
        "model.TeamAddUsers": {
            "type": "object",
            "required": [
                "member_uuids",
                "team_uuid"
            ],
            "properties": {
                "member_uuids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NewTeamMember"
                    }
                },
                "team_uuid": {
                    "type": "string"
                }
            }
        },
        "model.TeamCreate": {
            "type": "object",
            "required": [
                "member_uuids",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "member_uuids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NewTeamMember"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.TeamMember": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_day_of_week": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "is_manager": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "user_uuid": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "weekly_rate": {
                    "type": "integer"
                },
                "weekly_rate_name": {
                    "type": "string"
                },
                "work_session_status": {
                    "type": "string"
                }
            }
        },
        "model.TeamReadAll": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "team_members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TeamMember"
                    }
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "model.TeamUpdate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}